Gui Update 11/17/2018
Added Listeners to Enforce the text boxes will only revieve integers. Added Event Handlers for New, Exit, and Run. Run will call Supermarket with the contents of the text boxes as the arguments. New will populate the boxes with default values. Exit does.

I've noticed that while the simulation is running we lose control of the GUI. ("Stops Responding").